uint -> uint64

+++ bitset_benchmark_test.go
@@ -15,10 +15,10 @@ func BenchmarkSet(b *testing.B) {
 	b.StopTimer()
 	r := rand.New(rand.NewSource(0))
 	sz := 100000
-	s := New(uint(sz))
+	s := New(uint64(sz))
 	b.StartTimer()
 	for i := 0; i < b.N; i++ {
-		s.Set(uint(r.Int31n(int32(sz))))
+		s.Set(uint64(r.Int31n(int32(sz))))
 	}
 }
 
@@ -26,16 +26,16 @@ func BenchmarkGetTest(b *testing.B) {
 	b.StopTimer()
 	r := rand.New(rand.NewSource(0))
 	sz := 100000
-	s := New(uint(sz))
+	s := New(uint64(sz))
 	b.StartTimer()
 	for i := 0; i < b.N; i++ {
-		s.Test(uint(r.Int31n(int32(sz))))
+		s.Test(uint64(r.Int31n(int32(sz))))
 	}
 }
 
 func BenchmarkSetExpand(b *testing.B) {
 	b.StopTimer()
-	sz := uint(100000)
+	sz := uint64(100000)
 	b.StartTimer()
 	for i := 0; i < b.N; i++ {
 		var s BitSet
@@ -48,7 +48,7 @@ func BenchmarkCount(b *testing.B) {
 	b.StopTimer()
 	s := New(100000)
 	for i := 0; i < 100000; i += 100 {
-		s.Set(uint(i))
+		s.Set(uint64(i))
 	}
 	b.StartTimer()
 	for i := 0; i < b.N; i++ {
@@ -61,11 +61,11 @@ func BenchmarkIterate(b *testing.B) {
 	b.StopTimer()
 	s := New(10000)
 	for i := 0; i < 10000; i += 3 {
-		s.Set(uint(i))
+		s.Set(uint64(i))
 	}
 	b.StartTimer()
 	for j := 0; j < b.N; j++ {
-		c := uint(0)
+		c := uint64(0)
 		for i, e := s.NextSet(0); e; i, e = s.NextSet(i + 1) {
 			c++
 		}
@@ -77,11 +77,11 @@ func BenchmarkSparseIterate(b *testing.B) {
 	b.StopTimer()
 	s := New(100000)
 	for i := 0; i < 100000; i += 30 {
-		s.Set(uint(i))
+		s.Set(uint64(i))
 	}
 	b.StartTimer()
 	for j := 0; j < b.N; j++ {
-		c := uint(0)
+		c := uint64(0)
 		for i, e := s.NextSet(0); e; i, e = s.NextSet(i + 1) {
 			c++
 		}
@@ -93,7 +93,7 @@ func BenchmarkSparseIterate(b *testing.B) {
 func BenchmarkLemireCreate(b *testing.B) {
 	for i := 0; i < b.N; i++ {
 		bitmap := New(0) // we force dynamic memory allocation
-		for v := uint(0); v <= 100000000; v += 100 {
+		for v := uint64(0); v <= 100000000; v += 100 {
 			bitmap.Set(v)
 		}
 	}
@@ -103,11 +103,11 @@ func BenchmarkLemireCreate(b *testing.B) {
 // see http://lemire.me/blog/2016/09/22/swift-versus-java-the-bitset-performance-test/
 func BenchmarkLemireCount(b *testing.B) {
 	bitmap := New(100000000)
-	for v := uint(0); v <= 100000000; v += 100 {
+	for v := uint64(0); v <= 100000000; v += 100 {
 		bitmap.Set(v)
 	}
 	b.ResetTimer()
-	sum := uint(0)
+	sum := uint64(0)
 	for i := 0; i < b.N; i++ {
 		sum += bitmap.Count()
 	}
@@ -120,11 +120,11 @@ func BenchmarkLemireCount(b *testing.B) {
 // see http://lemire.me/blog/2016/09/22/swift-versus-java-the-bitset-performance-test/
 func BenchmarkLemireIterate(b *testing.B) {
 	bitmap := New(100000000)
-	for v := uint(0); v <= 100000000; v += 100 {
+	for v := uint64(0); v <= 100000000; v += 100 {
 		bitmap.Set(v)
 	}
 	b.ResetTimer()
-	sum := uint(0)
+	sum := uint64(0)
 	for i := 0; i < b.N; i++ {
 		for i, e := bitmap.NextSet(0); e; i, e = bitmap.NextSet(i + 1) {
 			sum++
